win: fix compiler warnings
- udp.c: Fix signedness warning in ternary operator by adding explicit cast to ensure consistent types
- util.c: Fix signed/unsigned comparison warning in loop condition by casting sizeof result to unsigned int
- win.h: Add missing braces in UV_ONCE_INIT to fix -Wmissing-braces warning

+++ src/win/udp.c
@@ -1158,7 +1158,7 @@ int uv__udp_try_send2(uv_udp_t* handle,
   for (i = 0; i < (int) count; i++) {
     r = uv_udp_try_send(handle, bufs[i], nbufs[i], addrs[i]);
     if (r < 0)
-      return i > 0 ? i : r;  /* Error if first packet, else send count. */
+      return i > 0 ? i : (int) r;  /* Error if first packet, else send count. */
   }
 
   return i;
